+++ example/lib/main.dart
@@ -25,13 +25,35 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  static const Color COLOR_LEFT = Colors.white;
-  static const Color COLOR_RIGHT = Color.fromARGB(255, 225, 255, 199);
-  static const Color COLOR_CAPTION = Color.fromARGB(255, 212, 234, 244);
-  static const EdgeInsets MARGIN_LEFT = EdgeInsets.only(top: 8, right: 48);
-  static const EdgeInsets MARGIN_LEFT_NEXT = EdgeInsets.only(top: 2, right: 48);
-  static const EdgeInsets MARGIN_RIGHT = EdgeInsets.only(top: 8, left: 48);
-  static const EdgeInsets MARGIN_RIGHT_NEXT = EdgeInsets.only(top: 2, left: 48);
+  static const BubbleStyle BUBBLE_CAPTION = BubbleStyle(
+    color: Color.fromARGB(255, 212, 234, 244),
+    margin: EdgeInsets.only(top: 8),
+    alignment: Alignment.topCenter,
+  );
+  static const BubbleStyle BUBBLE_SOMEBODY = BubbleStyle(
+    nip: BubbleNip.TOP_LEFT,
+    color: Colors.white,
+    margin: EdgeInsets.only(top: 8, right: 48),
+    alignment: Alignment.topLeft,
+  );
+  static const BubbleStyle BUBBLE_SOMEBODY_NEXT = BubbleStyle(
+    nip: BubbleNip.TOP_LEFT,
+    color: Colors.white,
+    margin: EdgeInsets.only(top: 2, right: 48),
+    alignment: Alignment.topLeft,
+  );
+  static const BubbleStyle BUBBLE_ME = BubbleStyle(
+    nip: BubbleNip.TOP_RIGHT,
+    color: Color.fromARGB(255, 225, 255, 199),
+    margin: EdgeInsets.only(top: 8, left: 48),
+    alignment: Alignment.topRight,
+  );
+  static const BubbleStyle BUBBLE_ME_NEXT = BubbleStyle(
+    nip: BubbleNip.TOP_RIGHT,
+    color: Color.fromARGB(255, 225, 255, 199),
+    margin: EdgeInsets.only(top: 2, left: 48),
+    alignment: Alignment.topRight,
+  );
 
   @override
   Widget build(BuildContext context) {
@@ -44,69 +66,33 @@ class _MyHomePageState extends State<MyHomePage> {
         child: ListView(
           padding: EdgeInsets.all(8),
           children: [
-            Container(
-              alignment: Alignment.topCenter,
-              child: Bubble(
-                nip: BubbleNip.NO,
-                color: COLOR_CAPTION,
-                child: Text('TODAY', style: TextStyle(fontSize: 10)),
-              ),
+            Bubble(
+              style: BUBBLE_CAPTION,
+              child: Text('TODAY', style: TextStyle(fontSize: 10)),
             ),
-            Container(
-              alignment: Alignment.topLeft,
-              child: Bubble(
-                nip: BubbleNip.TOP_LEFT,
-                color: COLOR_LEFT,
-                child: Text('Hi Jason. Sorry to bother you. I have a queston for you.'),
-                margin: MARGIN_LEFT,
-              ),
+            Bubble(
+              style: BUBBLE_SOMEBODY,
+              child: Text('Hi Jason. Sorry to bother you. I have a queston for you.'),
             ),
-            Container(
-              alignment: Alignment.topRight,
-              child: Bubble(
-                nip: BubbleNip.TOP_RIGHT,
-                color: COLOR_RIGHT,
-                child: Text('Whats\'up?'),
-                margin: MARGIN_RIGHT,
-              ),
+            Bubble(
+              style: BUBBLE_ME,
+              child: Text('Whats\'up?'),
             ),
-            Container(
-              alignment: Alignment.topLeft,
-              child: Bubble(
-                nip: BubbleNip.TOP_LEFT,
-                color: COLOR_LEFT,
-                child: Text('I\'ve been having a problem with my computer.'),
-                margin: MARGIN_LEFT,
-              ),
+            Bubble(
+              style: BUBBLE_SOMEBODY,
+              child: Text('I\'ve been having a problem with my computer.'),
             ),
-            Container(
-              alignment: Alignment.topLeft,
-              child: Bubble(
-                nip: BubbleNip.TOP_LEFT,
-                color: COLOR_LEFT,
-                showNip: false,
-                child: Text('Can you help me?'),
-                margin: MARGIN_LEFT_NEXT,
-              ),
+            Bubble(
+              style: BUBBLE_SOMEBODY_NEXT,
+              child: Text('Can you help me?'),
             ),
-            Container(
-              alignment: Alignment.topRight,
-              child: Bubble(
-                nip: BubbleNip.TOP_RIGHT,
-                color: COLOR_RIGHT,
-                child: Text('Ok'),
-                margin: MARGIN_RIGHT,
-              ),
+            Bubble(
+              style: BUBBLE_ME,
+              child: Text('Ok'),
             ),
-            Container(
-              alignment: Alignment.topRight,
-              child: Bubble(
-                nip: BubbleNip.TOP_RIGHT,
-                showNip: false,
-                color: COLOR_RIGHT,
-                child: Text('What\'s the problem?'),
-                margin: MARGIN_RIGHT_NEXT,
-              ),
+            Bubble(
+              style: BUBBLE_ME_NEXT,
+              child: Text('What\'s the problem?'),
             ),
           ],
         ),